kumo-screenshot-worker: split into modules
